feat: 恢复战斗玩家道具双数量步进
1. 复原共享玩家道具链、双数量上限、头插分配与四处caller直连。
2. 补齐定向测试、权威LST证据和第83项工作包进度。
3. 清理完整重编暴露的故事VM与SDL入口编译warning。

+++ tests/unit/world_map/legacy_world_story_vm_test_02.cpp
@@ -1429,7 +1429,7 @@ void test_clear_common_join_latch_and_yield_protocol(
         u16 previous_at_audio{};
         fixture.ports.audio_service_callback = [&] {
             offset_at_audio = fixture.context.instruction_offset;
-            previous_at_audio = fixture.state.previous_opcode;
+            previous_at_audio = static_cast<u16>(fixture.state.previous_opcode);
         };
 
         const auto result = fixture.step();
@@ -1463,7 +1463,7 @@ void test_clear_common_join_latch_and_yield_protocol(
     u16 previous_at_audio{};
     latched.ports.audio_service_callback = [&] {
         offset_at_audio = latched.context.instruction_offset;
-        previous_at_audio = latched.state.previous_opcode;
+        previous_at_audio = static_cast<u16>(latched.state.previous_opcode);
     };
 
     const auto latched_result = latched.step();
